requirements.txt
fastapi
uvicorn
python-multipart
Pillow
jinja2
python-doctr[torch]
numpy
